Add expected walking neighborhoods.

Add expected walking neighborhoods.

addNeighborhoodWalking(pump.subset = NULL, pump.select = NULL,
  vestry = FALSE, weighted = TRUE, polygon.method = "pearl.string",
  multi.core = FALSE, area = TRUE, path = NULL, path.color = NULL,
  path.width = 3, alpha.level = 0.25, polygon.type = "area",
  polygon.col = NULL, polygon.lwd = 2, ...)

Arguments

pump.subset
Numeric. Vector of numeric pump IDs to subset from the neighborhoods defined by pump.select. Negative selection possible. NULL uses all pumps in pump.select.

pump.select
Numeric. Numeric vector of pump IDs that define which pump neighborhoods to consider (i.e., specify the "population"). Negative selection possible. NULL selects all pumps.

vestry
Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.

weighted
Logical. TRUE computes shortest path weighted by road length. FALSE computes shortest path in terms of the number of nodes.

polygon.method
Character. Method of computing polygon vertices: "pearl.string" or "traveling.salesman".

multi.core
Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Windows, only multi.core = FALSE is available.

area
Logical. Area polygons.

path
Character. "expected" or "observed".

path.color
Character. Use a single color for all paths. NULL uses neighborhood colors defined by snowColors().

path.width
Numeric. Set width of paths.

alpha.level
Numeric. Alpha level transparency for area plot: a value in [0, 1].

polygon.type
Character. "perimeter" or "area".

polygon.col
Character.

polygon.lwd
Numeric.

...
Additional plotting parameters.

Note

This function is computationally intensive. On a single core of a 2.3 GHz Intel i7, plotting observed paths to PDF takes about 5 seconds while doing so for expected paths takes about 28 seconds. Using the parallel implementation on 4 physical (8 logical) cores, these times fall to about 4 and 11 seconds. Note that parallelization is currently only available on Linux and Mac, and that although some precautions are taken in R.app on macOS, the developers of the 'parallel' package, which neighborhoodWalking() uses, strongly discourage against using parallelization within a GUI or embedded environment. See vignette("parallel") for details.

Examples

# NOT RUN {
streetNameLocator("marshall street", zoom = TRUE, radius = 0.5)
addNeighborhoodWalking()
# }