Compute shortest walking distances or paths.

Compute shortest walking distances or paths.

nearestPump(pump.select = NULL, output = "distance", vestry = FALSE,
  weighted = TRUE, case.set = "observed", unit = "meter",
  multi.core = FALSE, time.unit = "second", walking.speed = 5)

Arguments

pump.select
Numeric. Pump candidates to consider. Default is NULL: all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for pumps; 1 to 14 for pumps.vestry. Negative selection allowed.

output
Character. "distance" or "path".

vestry
Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.

weighted
Logical. TRUE computes shortest path in terms of road length. FALSE computes shortest path in terms of the number of nodes.

case.set
Character. "observed", "expected", or "snow".

unit
Character. Unit of distance: "meter", "yard" or "native". "native" returns the map's native scale. Meaningful only when "weighted" is TRUE and "output" is "distance". See vignette("roads") for information on unit distances.

multi.core
Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Windows, only multi.core = FALSE is available.

time.unit
Character. "hour", "minute", or "second".

walking.speed
Numeric. Walking speed in km/hr.

Value

An R data frame or list of 'igraph' paths.

Note

Time is computed using distanceTime().