Compute Voronoi pump neighborhoods.

Group cases into neighborhoods using Voronoi tessellation.

neighborhoodVoronoi(pump.select = NULL, vestry = FALSE,
  case.location = "nominal", statistic = NULL,
  polygon.vertices = FALSE)

Arguments

pump.select
Numeric. Vector of numeric pump IDs to define pump neighborhoods (i.e., the "population"). Negative selection possible. NULL selects all pumps.

vestry
Logical. TRUE uses the 14 pumps from the Vestry report. FALSE uses the 13 in the original map.

case.location
Character. For observed = FALSE: "address" or "nominal". "nominal" is the x-y coordinate of regular.cases.

statistic
Character. NULL, the default, makes no summary computation. "address" computes the number of addresses in each selected pump neighborhood. "fatality" computes the number of fatalities in pump neighborhoods.

polygon.vertices
Logical. TRUE returns a list of x-y coordinates of the vertices of Voronoi cells. Useful for sp::point.in.polygon() as used in print.voronoi() method.

Value

An R list with 12 objects.

pump.id: vector of selected pumps

voronoi: output from deldir::deldir().

snow.colors: neighborhood color based on snowColors().

x.rng: range of x for plot.

y.rng: range of y for plot.

select.string: description of "pump.select" for plot title.

expected.data: expected neighborhood fatality counts, based on Voronoi cell area.

coordinates: polygon vertices of Voronoi cells.

statistic.data: observed neighborhood fatality counts.

pump.select: "pump.select" from neighborhoodVoronoi().

statistic: "statistic" from neighborhoodVoronoi().

vestry: "vestry" from neighborhoodVoronoi().

Examples

neighborhoodVoronoi()

#>  [1]   0   1  10  13   3  39 182  12  17  38   2   2   2

neighborhoodVoronoi(vestry = TRUE)

#>  [1]   0   1  10  15   3  36 182  11  17  40   2   2   2   0

neighborhoodVoronoi(pump.select = 6:7)

#> [1]  45 276

neighborhoodVoronoi(pump.select = -6)

#>  [1]   0   1  10  13  13 211  12  17  38   2   2   2

neighborhoodVoronoi(pump.select = -6, polygon.vertices = TRUE)

#> [[1]]
#>           x        y
#> 1  9.678367 18.72500
#> 2  3.390000 18.72500
#> 3  3.390000 16.32146
#> 4 10.296353 16.42222
#> 
#> [[2]]
#>           x        y
#> 1 12.542312 18.72500
#> 2  9.678367 18.72500
#> 3 10.296353 16.42222
#> 4 11.194356 15.85294
#> 
#> [[3]]
#>          x        y
#> 1 13.81843 18.72500
#> 2 12.54231 18.72500
#> 3 11.19436 15.85294
#> 4 11.79153 14.72931
#> 5 15.05211 14.26512
#> 
#> [[4]]
#>          x        y
#> 1 19.91200 18.72500
#> 2 13.81843 18.72500
#> 3 15.05211 14.26512
#> 4 16.55610 13.69434
#> 5 16.84189 13.74627
#> 6 19.91200 15.28076
#> 
#> [[5]]
#>           x         y
#> 1 11.791529 14.729307
#> 2 11.194356 15.852944
#> 3 10.296353 16.422219
#> 4  3.390000 16.321458
#> 5  3.390000  9.833668
#> 6  4.770556  9.876568
#> 7  8.649155 10.896535
#> 
#> [[6]]
#>           x         y
#> 1 16.556099 13.694337
#> 2 15.052111 14.265117
#> 3 11.791529 14.729307
#> 4  8.649155 10.896535
#> 5 11.753629  9.516815
#> 6 14.289421 10.156058
#> 
#> [[7]]
#>           x         y
#> 1 11.753629  9.516815
#> 2  8.649155 10.896535
#> 3  4.770556  9.876568
#> 4 12.496764  4.361363
#> 5 12.646096  4.696425
#> 
#> [[8]]
#>          x         y
#> 1 14.28942 10.156058
#> 2 11.75363  9.516815
#> 3 12.64610  4.696425
#> 4 15.59636  7.213220
#> 
#> [[9]]
#>          x         y
#> 1 16.84189 13.746269
#> 2 16.55610 13.694337
#> 3 14.28942 10.156058
#> 4 15.59636  7.213220
#> 5 18.17318  6.949869
#> 
#> [[10]]
#>          x         y
#> 1 19.91200 15.280758
#> 2 16.84189 13.746269
#> 3 18.17318  6.949869
#> 4 19.91200  5.871477
#> 
#> [[11]]
#>          x        y
#> 1 19.91200 5.871477
#> 2 18.17318 6.949869
#> 3 15.59636 7.213220
#> 4 12.64610 4.696425
#> 5 12.49676 4.361363
#> 6 12.48805 3.235000
#> 7 19.91200 3.235000
#> 
#> [[12]]
#>           x        y
#> 1  4.770556 9.876568
#> 2  3.390000 9.833668
#> 3  3.390000 3.235000
#> 4 12.488052 3.235000
#> 5 12.496764 4.361363
#> 

# coordinates for vertices also available in the returned object.
dat <- neighborhoodVoronoi(pump.select = -6)
dat$coordinates

#> [[1]]
#>           x        y
#> 1  9.678367 18.72500
#> 2  3.390000 18.72500
#> 3  3.390000 16.32146
#> 4 10.296353 16.42222
#> 
#> [[2]]
#>           x        y
#> 1 12.542312 18.72500
#> 2  9.678367 18.72500
#> 3 10.296353 16.42222
#> 4 11.194356 15.85294
#> 
#> [[3]]
#>          x        y
#> 1 13.81843 18.72500
#> 2 12.54231 18.72500
#> 3 11.19436 15.85294
#> 4 11.79153 14.72931
#> 5 15.05211 14.26512
#> 
#> [[4]]
#>          x        y
#> 1 19.91200 18.72500
#> 2 13.81843 18.72500
#> 3 15.05211 14.26512
#> 4 16.55610 13.69434
#> 5 16.84189 13.74627
#> 6 19.91200 15.28076
#> 
#> [[5]]
#>           x         y
#> 1 11.791529 14.729307
#> 2 11.194356 15.852944
#> 3 10.296353 16.422219
#> 4  3.390000 16.321458
#> 5  3.390000  9.833668
#> 6  4.770556  9.876568
#> 7  8.649155 10.896535
#> 
#> [[6]]
#>           x         y
#> 1 16.556099 13.694337
#> 2 15.052111 14.265117
#> 3 11.791529 14.729307
#> 4  8.649155 10.896535
#> 5 11.753629  9.516815
#> 6 14.289421 10.156058
#> 
#> [[7]]
#>           x         y
#> 1 11.753629  9.516815
#> 2  8.649155 10.896535
#> 3  4.770556  9.876568
#> 4 12.496764  4.361363
#> 5 12.646096  4.696425
#> 
#> [[8]]
#>          x         y
#> 1 14.28942 10.156058
#> 2 11.75363  9.516815
#> 3 12.64610  4.696425
#> 4 15.59636  7.213220
#> 
#> [[9]]
#>          x         y
#> 1 16.84189 13.746269
#> 2 16.55610 13.694337
#> 3 14.28942 10.156058
#> 4 15.59636  7.213220
#> 5 18.17318  6.949869
#> 
#> [[10]]
#>          x         y
#> 1 19.91200 15.280758
#> 2 16.84189 13.746269
#> 3 18.17318  6.949869
#> 4 19.91200  5.871477
#> 
#> [[11]]
#>          x        y
#> 1 19.91200 5.871477
#> 2 18.17318 6.949869
#> 3 15.59636 7.213220
#> 4 12.64610 4.696425
#> 5 12.49676 4.361363
#> 6 12.48805 3.235000
#> 7 19.91200 3.235000
#> 
#> [[12]]
#>           x        y
#> 1  4.770556 9.876568
#> 2  3.390000 9.833668
#> 3  3.390000 3.235000
#> 4 12.488052 3.235000
#> 5 12.496764 4.361363
#>