Compute walking path pump neighborhoods.

Group cases into neighborhoods based on walking distance.

neighborhoodWalking(pump.select = NULL, vestry = FALSE,
  weighted = TRUE, case.set = "observed", multi.core = FALSE)

Arguments

pump.select
Numeric. Vector of numeric pump IDs to define pump neighborhoods (i.e., the "population"). Negative selection possible. NULL selects all pumps. Note that you can't just select the pump on Adam and Eve Court (#2) because it's technically an isolate.

vestry
Logical. TRUE uses the 14 pumps from the Vestry report. FALSE uses the 13 in the original map.

weighted
Logical. TRUE computes shortest path weighted by road length. FALSE computes shortest path in terms of the number of nodes.

case.set
Character. "observed", "expected" or "snow". "snow" captures John Snow's annotation of the Broad Street pump neighborhood printed in the Vestry report version of the map.

multi.core
Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Windows, only multi.core = FALSE is available.

Value

An R list with 7 objects:

paths: list of paths to nearest or selected pump(s).

cases: list of cases by pump.

vestry: "vestry" from neighborhoodWalking().

observed: "observed" from neighborhoodWalking().

pump.select: "pump.select" from neighborhoodWalking().

cores: number of cores to use for parallel implementation.

metric: incremental metric used to find cut point on split road segments.

Note

This function is computationally intensive. On a single core of a 2.3 GHz Intel i7, plotting observed paths to PDF takes about 5 seconds while doing so for expected paths takes about 28 seconds. Using the parallel implementation on 4 physical (8 logical) cores, these times fall to about 4 and 11 seconds. Note that parallelization is currently only available on Linux and Mac, and that although some precautions are taken in R.app on macOS, the developers of the 'parallel' package, which neighborhoodWalking() uses, strongly discourage against using parallelization within a GUI or embedded environment. See vignette("parallel") for details.

Examples

# NOT RUN {
neighborhoodWalking()
neighborhoodWalking(pump.select = -6)
# }